requirements.txt
requests>=2.32.0
pandas
instagrapi
Pillow
python-dotenv
